==> 02_paths.R <==
#--------------------------------------------
## Set paths
#--------------------------------------------
paths <- list(
  cachePath = checkPath("cache", create = TRUE),
  modulePath = file.path("modules"),
  inputPath = file.path("inputs"),
  outputPath = file.path("outputs")
)
